Move state-abbrev to /lib
