+++ src/lib/index.js
@@ -1,6 +1,8 @@
 import React from 'react' // needed to get JSX in scope for text2paraJsx
 import Config from '../config'
 
+export { getStateFullName } from './state-abbrev'
+
 export const formatDate = (date) => {
   const monthAbbr = [
     'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun',
